# api/openapi.yaml
openapi: 3.0.3
info:
  title: Open API Kundenbeziehung - Swiss National Standard
  description: |
    Schweizer Standard für den branchenübergreifenden Austausch von Kundendaten zur Etablierung einer selbstbestimmten digitalen Kundenbeziehung.
    
    Diese API implementiert das vollständige Framework der Open API Kundenbeziehung mit dem universellen 10-Stufen-Referenzprozess und ermöglicht standardisierten Datenaustausch zwischen Finanzdienstleistern und anderen Ecosystemen.
    
    **Implementierte Module (Datenbausteine):**
    - Initialisierung (Step 1) - Service Discovery & Cookie Consent
    - Produktauswahl (Step 2) - Bedürfnisbefriedigung & Konfiguration  
    - Selbstdeklaration (Step 3) - FATCA/MIFID II Information
    - Basisdaten (Step 4) - Identität, Kontakt, Adresse
    - Erweiterte Daten (Step 5) - Risiko-/Potenzialermittlung
    - Identifikation (Step 6) - E-ID kompatibel, VideoIdent
    - Background Checks (Step 7) - KYC, AML, PEP, Sanctions
    - Vertragsabschluss (Step 8) - AGB-Akzeptanz
    - Signatur-Services (Step 9) - QES/AES/eSignatur
    - Metadaten-Management (Step 10) - Verarbeitung & Integration
    
    **Use Cases Implementiert:**
    - UC1: Kundenbeziehungseröffnung (Banking Account Opening)
    - UC2: Re-Identifikation (Cross-Provider Identity Reuse)
    - UC3: Altersverifikation (Privacy-preserving Age Verification)
    - UC4: CLM von EVV-Endkunden (External Wealth Management Lifecycle)
    
    **Sicherheitsstandards:**
...
    
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - mTLS: []
paths:
...
  /authorize:
    get:
      summary: OAuth 2.1 Authorization Endpoint
      description: |
        FAPI 2.0 compliant OAuth 2.1 authorization endpoint. Supports PAR (Pushed Authorization Requests)
        and requires PKCE for all authorization flows.
      operationId: authorize
      tags:
        - OAuth 2.1 / OIDC
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be 'code' for authorization code flow
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: OAuth client identifier
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Client redirect URI
        - name: scope
          in: query
          required: true
          schema:
            type: string
          description: Requested OAuth scopes
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: Client state parameter for CSRF protection
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge (S256)
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
          description: PKCE code challenge method
        - name: nonce
          in: query
          schema:
            type: string
          description: OpenID Connect nonce
        - name: request_uri
          in: query
          schema:
            type: string
          description: PAR request URI (urn:ietf:params:oauth:request_uri:*)
      responses:
        '302':
          description: Redirect to redirect_uri with authorization code
        '400':
          description: Invalid authorization request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security: []
...
  /userinfo:
    get:
      summary: OpenID Connect UserInfo Endpoint
      description: |
        Returns user information for the authenticated user. Supports DPoP token binding.
      operationId: userinfo
      tags:
        - OAuth 2.1 / OIDC
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [openid, profile]
        - DPoP: [openid, profile]
  /introspect:
    post:
      summary: OAuth 2.1 Token Introspection
      description: |
        RFC 7662 compliant token introspection endpoint for resource servers.
      operationId: introspect
      tags:
        - OAuth 2.1 / OIDC
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Token to introspect
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Hint about token type
      responses:
        '200':
          description: Token introspection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectionResponse'
        '400':
          description: Invalid introspection request
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - mTLS: []

  # PUSHED AUTHORIZATION REQUESTS (PAR)
...
  /consent:
    post:
      summary: Erstelle Consent-Request
      description: Initiiert einen Consent-Flow für Datenaustausch zwischen Institutionen
      operationId: createConsent
      tags:
        - Consent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
            examples:
              account_opening:
                summary: Kontoeröffnung
                value:
                  customerId: "sha256:a1b2c3d4e5f6..."
                  requestingInstitution: "CH-BANK-001"
                  providingInstitution: "CH-BANK-002"
                  dataCategories: ["basicData", "identification", "kycData"]
                  purpose: "accountOpening"
                  expiryDate: "2024-12-31T23:59:59Z"
                  customerContactMethod: "email"
      responses:
        '201':
          description: Consent erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /customer/check:
    post:
      summary: Kunden-Existenz prüfen
      description: Prüft ob ein Kunde bereits bei einer Institution identifiziert wurde (MVP Identifikation)
      operationId: checkCustomer
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCheckRequest'
            examples:
              basic_check:
                summary: Basis-Kundenprüfung
                value:
                  sharedCustomerHash: "sha256:abc123def456..."
                  basicData:
                    lastName: "Müller"
                    givenName: "Hans"
                    birthDate: "1985-03-15"
                    nationality: ["CH"]
      responses:
        '200':
          description: Kunde erfolgreich geprüft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /customer/data:
    post:
      summary: Kundendaten mit Consent abrufen
      description: Ruft spezifische Kundendatenmodule basierend auf gewährtem Consent ab
      operationId: getCustomerData
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDataRequest'
      responses:
        '200':
          description: Kundendaten erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  # PORTFOLIO & EVV SERVICES
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-Token mit Consent-Claims für API-Zugriff (FAPI 2.0 compliant)
    DPoP:
      type: http
      scheme: DPoP
      description: DPoP (Demonstrating Proof-of-Possession) für sender-constrained tokens
    mTLS:
      type: mutualTLS
      description: Mutual TLS für höchste Sicherheitsanforderungen (FAPI 2.0 compliant)
    private_key_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Private Key JWT client authentication (RFC 7523) - alternative zu mTLS
    RegistrationToken:
      type: http
      scheme: bearer
      description: Registration access token für dynamic client registration
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "INVALID_REQUEST"
            message: "Missing required field: customerId"
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or missing authentication token"
            timestamp: "2024-01-15T10:30:00Z"
    Forbidden:
      description: Zugriff verweigert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions for requested operation"
            timestamp: "2024-01-15T10:30:00Z"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Fehlercode
        message:
          type: string
          description: Fehlerbeschreibung
        timestamp:
          type: string
          format: date-time
          description: Zeitstempel des Fehlers
        details:
          type: object
          description: Zusätzliche Fehlerdetails
...
    ConsentRequest:
      type: object
      required:
        - customerId
        - requestingInstitution
        - dataCategories
        - purpose
        - expiryDate
      properties:
        customerId:
          type: string
          description: Eindeutige Kunden-ID (sharedCustomerHash)
          example: "sha256:abc123..."
        requestingInstitution:
          type: string
          description: Institution die Daten anfordert
          example: "CH-BANK-001"
        providingInstitution:
          type: string
          description: Institution die Daten bereitstellt
          example: "CH-BANK-002"
        dataCategories:
          type: array
          items:
            $ref: '#/components/schemas/DataCategory'
          description: Angeforderte Datenkategorien
        purpose:
          type: string
          enum: [accountOpening, creditAssessment, compliance, customerUpdate]
          description: Zweck der Datenverwendung
        expiryDate:
          type: string
          format: date-time
          description: Ablaufzeitpunkt des Consents
        customerContactMethod:
          type: string
          enum: [email, sms, app, postal]
          description: Bevorzugter Kontaktweg für Consent-Bestätigung
    ConsentResponse:
      type: object
      properties:
        consentId:
          type: string
          format: uuid
          description: Eindeutige Consent-ID
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        consentUrl:
          type: string
          format: uri
          description: URL für Kunden-Consent-Interface
        qrCode:
          type: string
          description: Base64-kodierter QR-Code für mobilen Consent
        expiryDate:
          type: string
          format: date-time
...
    DataCategory:
      type: string
      enum: 
        - basicData
        - identification
        - contactInformation
        - addressData
        - kycData
        - riskProfile
        - complianceData
        - extendedData

    # CUSTOMER DATA SCHEMAS
    CustomerCheckRequest:
      type: object
      required:
        - sharedCustomerHash
        - basicData
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash der Grunddaten für Matching
          example: "sha256:a1b2c3d4..."
        basicData:
          $ref: '#/components/schemas/BasicCustomerData'
    CustomerCheckResponse:
      type: object
      properties:
        match:
          type: boolean
          description: Ob ein Kunde gefunden wurde
        identificationDate:
          type: string
          format: date-time
          description: Datum der letzten Identifikation
        levelOfAssurance:
          type: string
          enum: [low, substantial, high]
          description: Sicherheitsniveau der Identifikation
        validUntil:
          type: string
          format: date-time
          description: Gültigkeit der Identifikation
...
    BasicCustomerData:
      type: object
      required:
        - lastName
        - givenName
        - birthDate
        - nationality
      properties:
        lastName:
          type: string
          description: Nachname
          example: "Müller"
          maxLength: 100
        givenName:
          type: string
          description: Vorname(n)
          example: "Hans Peter"
          maxLength: 100
        middleName:
          type: string
          description: Weitere Vornamen
          maxLength: 100
        title:
          type: string
          description: Titel
          example: "Dr."
          maxLength: 50
        birthDate:
          type: string
          format: date
          description: Geburtsdatum
          example: "1985-03-15"
        birthPlace:
          type: string
          description: Geburtsort
          example: "Zürich"
          maxLength: 100
        nationality:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{2}$'
          description: Nationalität(en) (ISO 3166-1)
          example: ["CH", "DE"]
          minItems: 1
          maxItems: 5
        gender:
          type: string
          enum: [male, female, other, unknown]
        maritalStatus:
          type: string
          enum: [single, married, divorced, widowed, registered_partnership]
        language:
          type: string
          description: Bevorzugte Sprache (ISO 639-1)
          example: "de"
          pattern: '^[a-z]{2}$'
...
    CustomerDataRequest:
      type: object
      required:
        - sharedCustomerHash
        - requestedModules
        - consentToken
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash des Kunden
        requestedModules:
          type: array
          items:
            type: string
            enum: [basisdaten_module, erweiterte_daten_module, identifikation_module, background_checks_module]
          description: Angeforderte Datenbausteine
        consentToken:
          type: string
          description: JWT-Token mit Consent-Nachweis
        purpose:
          type: string
          enum: [kundenbeziehungseroffnung, re_identification, compliance_update]
    CustomerDataResponse:
      type: object
      properties:
        availableModules:
          type: array
          items:
            type: string
        moduleData:
          type: object
          description: Bereitgestellte Datenbausteine
        verificationStatus:
          type: object
        lastUpdated:
          type: string
          format: date-time
...
    OAuthError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - server_error
            - temporarily_unavailable
        error_description:
          type: string
          description: Human-readable error description
        error_uri:
          type: string
          format: uri
          description: URI to error documentation
        timestamp:
          type: string
          format: date-time
...
    UserInfo:
      type: object
      required:
        - sub
      properties:
        sub:
          type: string
          description: Subject identifier
        preferred_username:
          type: string
          description: Preferred username
        name:
          type: string
          description: Full name
        given_name:
          type: string
          description: Given name
        family_name:
          type: string
          description: Family name
        email:
          type: string
          format: email
          description: Email address
        email_verified:
          type: boolean
          description: Whether email is verified
        institution_id:
          type: string
          description: Associated institution identifier
        updated_at:
          type: integer
          description: Time of last update (Unix timestamp)
    IntrospectionResponse:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          description: Whether the token is active
        client_id:
          type: string
          description: Client identifier
        scope:
          type: string
          description: Authorized scopes
        sub:
          type: string
          description: Subject of the token
        aud:
          type: string
          description: Intended audience
        iss:
          type: string
          description: Token issuer
        exp:
          type: integer
          description: Expiration time (Unix timestamp)
        iat:
          type: integer
          description: Issued at time (Unix timestamp)
        token_type:
          type: string
          enum: [Bearer, DPoP]
          description: Token type

    # Client Management Schemas
...